refactor: move component to a specific wallet folder
