+++ src/components/wallet/mnemonicPopUp.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import mnemonicWord from '@/components/mnemonicWord.vue'
+import mnemonicWord from '@/components/wallet/mnemonicWord.vue'
 
 export default {
   name: 'mnemonicPopUp',
@@ -25,7 +25,7 @@ export default {
 
 <template>
   <div
-    class="relative bg-background-gray rounded-lg w-2/3 h-2/3 p-6 md:p-8 lg:p-10 flex flex-col"
+    class="relative bg-background-gray rounded-lg w-2/3 h-2/3 p-6 md:p-8 lg:p-10 flex flex-col z-10"
     @keydown.esc="$emit('close')"
     tabindex="0"
   >
