+++ src/components/wallet/mnemonicWord.vue
